[DEV-4963] fix imports
